+++ CHANGELOG.md
@@ -8,7 +8,7 @@
 - Offer to copy non-HTTP links to the clipboard instead of trying to open them in webxdc source code link and inside of html emails.
 
 ### Fixed
-
+- fix duplicated calling of 'open-url'
 
 <a id="1_40_3"></a>